+++ backend/collator/src/organization.rs
@@ -167,11 +167,11 @@ WHERE
         r#"
 SELECT
   "region" AS "state!: String",
-  SUM("total_users") AS "unique_users: i64",
-  SUM("new_users") AS "new_users: i64",
-  COALESCE(SUM("total_users"), 0) - COALESCE(SUM("new_users"), 0) AS "returning_users: i64",
-  SUM("views") AS "total_pageviews: i64",
-  SUM("sessions") AS "unique_pageviews: i64",
+  SUM("total_users")::bigint AS "unique_users: i64",
+  SUM("new_users")::bigint AS "new_users: i64",
+  COALESCE(SUM("total_users")::bigint, 0) - COALESCE(SUM("new_users")::bigint, 0) AS "returning_users: i64",
+  SUM("views")::bigint AS "total_pageviews: i64",
+  SUM("sessions")::bigint AS "unique_pageviews: i64",
   AVG("engagement_duration") AS "average_time: f64"
 FROM c_analytics_cache
 WHERE "begin" = $1 AND "end" = $2 AND "partner" = $3 AND c_opportunity_by_uid_is_status("opportunity", $4)
@@ -244,11 +244,11 @@ GROUP BY "region"
             r#"
 SELECT
   "city" AS "region!: String",
-  SUM("total_users") AS "unique_users: i64",
-  SUM("new_users") AS "new_users: i64",
-  COALESCE(SUM("total_users"), 0) - COALESCE(SUM("new_users"), 0) AS "returning_users: i64",
-  SUM("views") AS "total_pageviews: i64",
-  SUM("sessions") AS "unique_pageviews: i64",
+  SUM("total_users")::bigint AS "unique_users: i64",
+  SUM("new_users")::bigint AS "new_users: i64",
+  COALESCE(SUM("total_users")::bigint, 0) - COALESCE(SUM("new_users")::bigint, 0) AS "returning_users: i64",
+  SUM("views")::bigint AS "total_pageviews: i64",
+  SUM("sessions")::bigint AS "unique_pageviews: i64",
   AVG("engagement_duration") AS "average_time: f64"
 FROM c_analytics_cache
 WHERE "begin" = $1 AND "end" = $2 AND "region" = $3 AND "partner" = $4 AND c_opportunity_by_uid_is_status("opportunity", $5)
@@ -366,11 +366,11 @@ GROUP BY "city"
         r#"
 SELECT
   "device_category" AS "device_category!: String",
-  SUM("total_users") AS "unique_users: i64",
-  SUM("new_users") AS "new_users: i64",
-  COALESCE(SUM("total_users"), 0) - COALESCE(SUM("new_users"), 0) AS "returning_users: i64",
-  SUM("views") AS "total_pageviews: i64",
-  SUM("sessions") AS "unique_pageviews: i64",
+  SUM("total_users")::bigint AS "unique_users: i64",
+  SUM("new_users")::bigint AS "new_users: i64",
+  COALESCE(SUM("total_users")::bigint, 0) - COALESCE(SUM("new_users")::bigint, 0) AS "returning_users: i64",
+  SUM("views")::bigint AS "total_pageviews: i64",
+  SUM("sessions")::bigint AS "unique_pageviews: i64",
   AVG("engagement_duration") AS "average_time: f64"
 FROM c_analytics_cache
 WHERE "begin" = $1 AND "end" = $2 AND "partner" = $3 AND c_opportunity_by_uid_is_status("opportunity", $4)
@@ -443,10 +443,10 @@ GROUP BY "device_category"
         r#"
 SELECT
   "date" AS "date!: DateTime<FixedOffset>",
-  SUM("views") AS "views: i64",
-  SUM("sessions") AS "unique: i64",
-  SUM("new_users") AS "new: i64",
-  SUM("sessions") - SUM("new_users") AS "returning: i64",
+  SUM("views")::bigint AS "views: i64",
+  SUM("sessions")::bigint AS "unique: i64",
+  SUM("new_users")::bigint AS "new: i64",
+  SUM("sessions")::bigint - SUM("new_users")::bigint AS "returning: i64",
   (
     SELECT COUNT(*)
     FROM c_log INNER JOIN c_opportunity ON c_log."object" = (c_opportunity.exterior->>'uid')::uuid
@@ -505,7 +505,7 @@ GROUP BY "date"
             ],
             data: sqlx::query!(
                 r#"
-SELECT "session_channel_group" AS "group!", SUM("views") AS "count: i64"
+SELECT "session_channel_group" AS "group!", SUM("views")::bigint AS "count: i64"
 FROM c_analytics_cache
 WHERE "partner" = $1 AND "date" >= $2 AND "date" < $3 AND c_opportunity_by_uid_is_status("opportunity", $4)
 GROUP BY "session_channel_group"
@@ -529,11 +529,11 @@ ORDER BY "session_channel_group"
 SELECT
   "page_referrer" AS "page_referrer!",
   "session_channel_group" AS "type_!",
-  SUM("total_users") AS "unique_users: i64",
-  SUM("new_users") AS "new_users: i64",
-  COALESCE(SUM("total_users"), 0) - COALESCE(SUM("new_users"), 0) AS "returning_users: i64",
-  SUM("views") AS "total_pageviews: i64",
-  SUM("sessions") AS "unique_pageviews: i64",
+  SUM("total_users")::bigint AS "unique_users: i64",
+  SUM("new_users")::bigint AS "new_users: i64",
+  COALESCE(SUM("total_users")::bigint, 0) - COALESCE(SUM("new_users")::bigint, 0) AS "returning_users: i64",
+  SUM("views")::bigint AS "total_pageviews: i64",
+  SUM("sessions")::bigint AS "unique_pageviews: i64",
   AVG("engagement_duration") AS "average_time: f64"
 FROM c_analytics_cache
 WHERE "begin" = $1 AND "end" = $2 AND "partner" = $3 AND c_opportunity_by_uid_is_status("opportunity", $4)
